requirements.txt
pywin32
Pillow
pypdf
